requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.0
python-multipart>=0.0.20
gpxpy>=1.6.2
python-dotenv>=1.1.0
pydantic>=2.11.1